+++ R/AFM.histogram.R
@@ -22,7 +22,7 @@
 #'
 #' @seealso \code{\link{plot.AFMdata}}
 #'
-#' @importFrom ggplot2 ggplot aes geom_histogram geom_density after_stat
+#' @importFrom ggplot2 ggplot aes geom_histogram geom_density after_stat 
 #' @importFrom graphics hist
 #'
 #' @export