+++ tests/testthat/test-hhcf.R
@@ -1,18 +1,21 @@
+filename = AFM.getSampleImages(type='tiff')
+a = AFM.import(filename)
+a = AFM.flatten(a)
+
 test_that("Height height correlation function: check roughness", {
-  filename = AFM.getSampleImages(type='tiff')
-  a = AFM.import(filename)
-  a = AFM.flatten(a)
   h1 = AFM.hhcf(a, numIterations = 1e5, degRes=1000, dataOnly = TRUE)
   # AFM.hhcf(a, numIterations = 4e4, degRes=1000, addFit=TRUE)
 
   expect_equal(mean(h1$g) / (2*AFM.math.params(a)$Rq^2), 1, tolerance = 2e-2)
 })
 
-test_that("HHCF should output fit parameters", {
-  filename = AFM.getSampleImages(type='tiff')
-  a = AFM.import(filename)
-  h1 = AFM.hhcf(a, numIterations = 1e5, degRes=1000, allResults = TRUE)
-  fit = h1$fitParams
-  expect_equal(fit$xi,40.0, tolerance = 0.02)
-  expect_equal(fit$H, 1.73, tolerance = 0.05)
+
+
+test_that("Height height correlation function: reproducibility", {
+  h1 = AFM.hhcf(a, numIterations = 1e4, degRes=1000, randomSeed = 45792231, allResults = TRUE)
+  # AFM.hhcf(a, numIterations = 4e4, degRes=1000, addFit=TRUE)
+  m1 =  h1$fitParams
+  h1 = AFM.hhcf(a, numIterations = 1e4, degRes=1000, randomSeed = 45792231, allResults = TRUE)
+  m2 =  h1$fitParams
+  expect_equal(m1,m2)
 })